LH #81
Added support for userInfo dictionary to the DPContributionNavigationDescriptor protocol. This makes it much easier to interact with couch elements in the left-hand nav. 

Also fixed the bug that was causing the admin plugin to fail. It isn't fast and needs more work but we'll get to that later. Once things are stable and feature complete, we'll need to do some instrumentation to improve performance.